Brunton Widget can now accept to datasets and one. Some structural changes regarding testing
