+++ Modeling/TF/TF_Functions/Test_predictions.py
@@ -13,7 +13,8 @@ from tqdm import trange
 import copy
 from time import sleep
 
-from Modeling.load_and_normalize import denormalize_df, load_data, load_normalization_info, normalize_df, denormalize_numpy_array
+from Modeling.load_and_normalize import denormalize_df, get_paths_to_datafiles, load_data, load_normalization_info, \
+    normalize_df, denormalize_numpy_array
 from Modeling.TF.TF_Functions.Network import get_internal_states, load_internal_states
 
 # Import functions from PyQt5 module (creating GUI)
@@ -28,6 +29,8 @@ from matplotlib.figure import Figure
 # This import mus go before pyplot so also before our scripts
 from matplotlib import use, get_backend
 # Use Agg if not in scientific mode of Pycharm
+from .Initialization import get_net_and_norm_info
+
 if get_backend() != 'module://backend_interagg':
     use('Agg')
 
@@ -56,12 +59,12 @@ cdict = {'red':   ((0.0,  0.22, 0.0),
 
 cmap = colors.LinearSegmentedColormap('custom', cdict)
 
-def open_loop_prediction_experiment(net_for_inference,
-                                    net_for_inference_info,
-                                    dataset=None,
-                                    normalization_info=None,
-                                    experiment_length=None
-                                    ):
+def get_predictions_TF(net_for_inference,
+                       net_for_inference_info,
+                       dataset=None,
+                       normalization_info=None,
+                       experiment_length=None
+                       ):
 
     # How far NN should predict
     max_horizon = 10
@@ -118,260 +121,25 @@ def open_loop_prediction_experiment(net_for_inference,
     # time_axis is a time axis for ground truth
     return ground_truth[:-max_horizon, :], net_outputs[:, :-max_horizon, :], time_axis[:-max_horizon]
 
-from PyQt5.QtWidgets import QApplication
 
-def run_test_gui(inputs, outputs, ground_truth, net_outputs, time_axis):
-    # Creat an instance of PyQt5 application
-    # Every PyQt5 application has to contain this line
-    app = QApplication(sys.argv)
-    # Create an instance of the GUI window.
-    window = MainWindow(inputs, outputs, ground_truth, net_outputs, time_axis)
-    window.show()
-    # Next line hands the control over to Python GUI
-    app.exec_()
-
-# Class implementing the main window of CartPole GUI
-class MainWindow(QMainWindow):
-
-    def __init__(self,
-                 inputs, outputs, ground_truth, net_outputs, time_axis,
-                 *args, **kwargs):
-        super(MainWindow, self).__init__(*args, **kwargs)
-
-        self.inputs = inputs
-        self.outputs = outputs
-        self.ground_truth = ground_truth
-        self.net_outputs = net_outputs
-        self.time_axis = time_axis
-
-        self.max_horizon = self.net_outputs.shape[0]
-        self.horizon = self.max_horizon//2
-
-        self.show_all = False
-        self.downsample = False
-        self.current_point_at_timeaxis = (self.time_axis.shape[0]-self.max_horizon)//2
-        self.feature_to_display = outputs[0]
-
-        # region - Create container for top level layout
-        layout = QVBoxLayout()
-        # endregion
-
-        # region - Change geometry of the main window
-        self.setGeometry(300, 300, 2500, 1000)
-        # endregion
-
-        # region - Feature selection
-
-        # endregion
-
-        # region - Matplotlib figures (CartPole drawing and Slider)
-        # Draw Figure
-        self.fig = Figure(figsize=(25, 10))  # Regulates the size of Figure in inches, before scaling to window size.
-        self.canvas = FigureCanvas(self.fig)
-        self.fig.Ax = self.canvas.figure.add_subplot(111)
-        self.redraw_canvas()
-
-        # Attach figure to the layout
-        lf = QVBoxLayout()
-        lf.addWidget(self.canvas)
-        layout.addLayout(lf)
-
-        # endregion
-
-        l_sl = QVBoxLayout()
-
-        # region - Slider position
-        l_sl.addWidget(QLabel('"Current" point in time:'))
-        self.sl_p = QSlider(Qt.Horizontal)
-        self.sl_p.setMinimum(0)
-        self.sl_p.setMaximum(self.time_axis.shape[0]-self.max_horizon)
-        self.sl_p.setValue((self.time_axis.shape[0]-self.max_horizon)//2)
-        self.sl_p.setTickPosition(QSlider.TicksBelow)
-        # self.sl_p.setTickInterval(5)
-
-        l_sl.addWidget(self.sl_p)
-        self.sl_p.valueChanged.connect(self.slider_position_f)
-        # endregion
-
-        # region - Slider horizon
-        l_sl.addWidget(QLabel('Prediction horizon:'))
-        self.sl_h = QSlider(Qt.Horizontal)
-        self.sl_h.setMinimum(0)
-        self.sl_h.setMaximum(self.max_horizon)
-        self.sl_h.setValue(self.max_horizon//2)
-        self.sl_h.setTickPosition(QSlider.TicksBelow)
-        # self.sl_h.setTickInterval(5)
-        # endregion
-
-        l_sl.addWidget(self.sl_h)
-        self.sl_h.valueChanged.connect(self.slider_horizon_f)
-
-        layout.addLayout(l_sl)
+def get_data_for_gui_TF(a):
+    # Create a copy of the network suitable for inference (stateful and with sequence length one)
+    net_for_inference, net_for_inference_info, normalization_info = \
+        get_net_and_norm_info(a, time_series_length=1,
+                              batch_size=1, stateful=True)
 
 
-        # region - Make strip of layout for checkboxes and compobox
-        l_cb = QHBoxLayout()
+    # region In either case testing is done on a data collected offline
+    paths_to_datafiles_test = get_paths_to_datafiles(a.test_files)
+    test_dfs = load_data(paths_to_datafiles_test)
+    test_dfs_norm = normalize_df(test_dfs, normalization_info)
+    test_set = Dataset(test_dfs_norm, a, shuffle=False,
+                       inputs=net_for_inference_info.inputs, outputs=net_for_inference_info.outputs)
 
-        # region -- Checkbox: Show all
-        self.cb_show_all = QCheckBox('Show all', self)
-        if self.show_all:
-            self.cb_show_all.toggle()
-        self.cb_show_all.toggled.connect(self.cb_show_all_f)
-        l_cb.addWidget(self.cb_show_all)
-        # endregion
+    ground_truth, net_outputs, time_axis = \
+        get_predictions_TF(net_for_inference, net_for_inference_info,
+                           test_set, normalization_info,
+                           experiment_length=a.test_len)
 
-        # region -- Checkbox: Save/don't save experiment recording
-        self.cb_downsample = QCheckBox('Downsample predictions (X2)', self)
-        if self.downsample:
-            self.cb_downsample.toggle()
-        self.cb_downsample.toggled.connect(self.cb_downsample_f)
-        l_cb.addWidget(self.cb_downsample)
-        # endregion
-
-        l_cb.addStretch(1)
-
-        # region -- Combobox: Select feature to plot
-        l_cb.addWidget(QLabel('Feature to plot:'))
-        self.cb_select_feature = QComboBox()
-        self.cb_select_feature.addItems(outputs)
-        self.cb_select_feature.currentIndexChanged.connect(self.cb_select_feature_f)
-        self.cb_select_feature.setCurrentText(outputs[0])
-        l_cb.addWidget(self.cb_select_feature)
-
-        # region - Add checkboxes to layout
-        layout.addLayout(l_cb)
-        # endregion
-
-        # endregion
-
-        # region - QUIT button
-        bq = QPushButton("QUIT")
-        bq.pressed.connect(self.quit_application)
-        lb = QVBoxLayout()  # Layout for buttons
-        lb.addWidget(bq)
-        layout.addLayout(lb)
-        # endregion
-
-        # region - Create an instance of a GUI window
-        w = QWidget()
-        w.setLayout(layout)
-        self.setCentralWidget(w)
-        self.show()
-        self.setWindowTitle('Testing TF model')
-
-        # endregion
-
-
-    def slider_position_f(self, value):
-        self.current_point_at_timeaxis = int(value)
-
-        self.redraw_canvas()
-
-    def slider_horizon_f(self, value):
-        self.horizon = int(value)
-
-        self.redraw_canvas()
-
-    def cb_show_all_f(self, state):
-        if state:
-            self.show_all = True
-        else:
-            self.show_all = False
-
-        self.redraw_canvas()
-
-    def cb_downsample_f(self, state):
-        if state:
-            self.downsample = True
-        else:
-            self.downsample = False
-
-        self.redraw_canvas()
-
-    def cb_select_feature_f(self):
-        self.feature_to_display = self.cb_select_feature.currentText()
-        self.redraw_canvas()
-
-    # The actions which has to be taken to properly terminate the application
-    # The method is evoked after QUIT button is pressed
-    # TODO: Can we connect it somehow also the the default cross closing the application?
-    def quit_application(self):
-        # Closes the GUI window
-        self.close()
-        # The standard command
-        # It seems however not to be working by its own
-        # I don't know how it works
-        QApplication.quit()
-
-
-    def redraw_canvas(self):
-
-        self.fig.Ax.clear()
-
-        brunton_widget(self.inputs, self.outputs, self.ground_truth, self.net_outputs, self.time_axis,
-                       axs=self.fig.Ax,
-                       current_point_at_timeaxis=self.current_point_at_timeaxis,
-                       feature_to_display=self.feature_to_display,
-                       max_horizon=self.max_horizon,
-                       horizon=self.horizon,
-                       show_all=self.show_all,
-                       downsample=self.downsample)
-
-        self.canvas.draw()
-
-
-
-def brunton_widget(inputs, outputs, ground_truth, net_outputs, time_axis, axs=None,
-                   current_point_at_timeaxis=None,
-                   feature_to_display=None,
-                   max_horizon=10, horizon=None,
-                   show_all=True,
-                   downsample=False):
-
-    # Start at should be done bu widget (slider)
-    if current_point_at_timeaxis is None:
-        current_point_at_timeaxis = ground_truth.shape[0]//2
-
-    if feature_to_display is None:
-        feature_to_display = 's.angle.cos'
-
-    if horizon is None:
-        horizon = max_horizon
-
-    feature_idx = inputs.index(feature_to_display)
-    target_idx = outputs.index(feature_to_display)
-
-    # Brunton Plot
-    if axs is None:
-        fig, axs = plt.subplots(1, 1, figsize=(18, 10), sharex=True)
-
-    axs.plot(time_axis, ground_truth[:, feature_idx], 'k:', markersize=12, label='Ground Truth')
-    y_lim = axs.get_ylim()
-    prediction_distance = []
-    axs.set_ylabel(feature_to_display, fontsize=18)
-    axs.set_xlabel('Time [s]', fontsize=18)
-    for i in range(horizon):
-
-        if not show_all:
-            axs.plot(time_axis[current_point_at_timeaxis], ground_truth[current_point_at_timeaxis, feature_idx],
-                     'g.', markersize=16, label='Start')
-            prediction_distance.append(net_outputs[i, current_point_at_timeaxis, target_idx])
-            if downsample:
-                if (i % 2) == 0:
-                    continue
-            axs.plot(time_axis[current_point_at_timeaxis+i+1], prediction_distance[i],
-                        c=cmap(float(i)/max_horizon),
-                        marker='.')
-
-        else:
-            prediction_distance.append(net_outputs[i, :-(i+1), target_idx])
-            if downsample:
-                if (i % 2) == 0:
-                    continue
-            axs.plot(time_axis[i+1:], prediction_distance[i],
-                        c=cmap(float(i)/max_horizon),
-                        marker='.', linestyle = '')
-
-    axs.set_ylim(y_lim)
-
-    plt.show()
+    return net_for_inference_info.inputs, net_for_inference_info.outputs, net_for_inference_info.net_full_name,\
+           ground_truth, net_outputs, time_axis
